+++ src/main/java/com/example/demo/member/controller/LoginController.java
@@ -1,22 +1,12 @@
 package com.example.demo.member.controller;
 
-import com.example.demo.member.jwt.JwtTokenProvider;
+import com.example.demo.security.jwt.JwtTokenProvider;
 import com.example.demo.member.mapper.UserMapper;
 import com.example.demo.member.service.impl.UserServiceImpl;
-import com.example.demo.member.vo.MessageVO;
-import com.example.demo.member.vo.UserVO;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-
-import javax.servlet.http.HttpServletRequest;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 @RequiredArgsConstructor
 @Controller
