+++ src/main/java/com/example/demo/security/jwt/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.example.demo.member.jwt;
+package com.example.demo.security.jwt;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
